Consolidate error helpers into main errors package and remove confusing panic.go
- Move NewInternalError, NewConfigurationError, ValidateCondition, ValidateNotNil to errors.go
- Remove confusing panic.go file that suggested panic usage when policy is no-panic
- Consolidate tests into main errors_test.go file
- Maintain all functionality while improving package organization
- Eliminate misleading filename that contradicted no-panic policy
